Move MCP UI capability tests to sibling module
